requirements.txt
bert_score==0.3.11
boto3==1.21.42
botocore==1.24.42
colorlog==6.6.0
fairseq==0.10.2
ftfy==6.1.1
h5py==3.6.0
jieba==0.42.1
nltk==3.7
numpy==1.20.2
pandas==1.3.5
prettytable==3.2.0
pytest==7.1.1
pythainlp==3.0.5
PyYAML==6.0
regex==2022.3.15
requests==2.24.0
sacremoses==0.0.49
scikit_learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
six==1.15.0
spacy==3.2.4
tensorboardX==2.5
tensorboard==2.8.0
torch==1.8.1
tqdm==4.51.0
transformers==4.18.0
